restructures folders to allow scaling of scraping across topics
